Update unit tests for Private route Total of 7 unit tests for Private.js.
First three tests check for null, valid and
invalid authentication tokens respectively. They
pass as expected.

Test 4 checks for when the axios.get call throws
an error. However, the test always throws an error
and causes all tests, including itself, to not
run. Unsure if this is the right way to test for
this scenario. Have skipped it for now.

Last three tests check for res, res.data and
res.data.ok being null. Test 5 and 6 throw errors
as they are not accounted for in original code and
causes all tests, including themselves, to not
run. Unsure if this is the right way to test for
these scenarios. Have skipped them for now.